Move module_interface and modules to app/api/api_v1
